R2-07 R4e: the Results lab built its renderer -mthumb, not -marm
Reported as an optimization; it is an instrument defect, and the
correction matters more than the number.

Chasing R4c's fighter cost, the Results census put __aeabi_lmul -- the
64-bit multiply helper -- at 20,935,676 cycles, 7.79% of the frame, third
behind the idle spin and the fighter root, out of 86 bytes of code.
objdump put 31 of its 79 call sites in five functions that all
disassembled to Thumb with zero SMULL between them.

ARMv5TE Thumb has no SMULL. It has no 32x32->64 multiply at all, so every
`(s64)a * b` in a Thumb function is a `bl __aeabi_lmul` however the
operands are typed. The 20.12 fixed-point matrix and vertex math is
exactly that shape, per vertex.

The cause was one line:

  ifeq ($(NDS_DEV_SCENE_HARNESS_ID),163)     # battle only
  nds_renderer.o: CFLAGS += -marm
  endif

results_playable is harness 164. It was added to the tick-HUD Makefile
block under a comment promising it must differ from the tick-HUD ROM "in
the scene it boots and in NOTHING else", and this rule -- keyed on the
harness ID rather than the target -- silently broke that promise. Every
Results number in this campaign (R0c, R0d, R0e, R2a, R2b, R4b, R4d) was
therefore measured on a Thumb-compiled renderer while battle measured an
ARM-compiled one. Those deltas hold against each other, since all arms
shared the defect; their absolute costs were inflated.

Keyed now on the latency surfaces (NDS_ARM_RENDERER_HARNESS_IDS := 163
164) so the next latency ROM cannot inherit the same trap.

  VBlanks/tic   4.03 -> 3.04
  ticks/tic     2,254,765 -> 1,701,577   -553,188 (-24.5%)
  FTR/tic       1,710,498 -> 1,449,776   -260,722 (-15.2%)
  vs 1.12M      2.01x -> 1.52x

Guest picture PIXEL-IDENTICAL against HEAD: 240,000 guest-viewport pixels
at Results tic 160, 0 differing, max channel delta 0. Latest green,
including battle_playable_realtime.

One approach was built and withdrawn. Before the Makefile line was found,
the same win was chased with __attribute__((target("arm"))) on the five
hot functions behind an NDS_R2_ARM_MULTIPLY flag. It worked -- 3.11
VBlanks, 1,743,591 -- but it is a hand-rolled partial duplicate of a
compile flag the target should already have had, covering five functions
instead of the translation unit, and it measured 42,014 ticks WORSE than
fixing the flag. Flag, macro and all eight attribute sites deleted.

What survives from it is one line of C: ndsRendererTransformVertex20p12
declared x, y, z as s64 when the source fields are s16 and the matrix is
s32, so every product fit in 47 bits and never needed the 64x64 path.
They are s32 now -- same value, one SMULL instead of a widened sequence.

The battle path was measured and is unaffected: it already had -marm.
Both battle arms produced a byte-identical loadable ARM9 image and
identical 128-sample percentiles. That is how the flag was found to be
inert there -- the tick-HUD sampler reported the same build sha twice,
and the .o files differed while objcopy -O binary output did not. No
battle P95 in the campaign moves.

+++ src/nds/nds_renderer.c
@@ -5123,9 +5123,13 @@ void ndsRendererTransformVertex20p12(const NDSRendererMatrix20p12 *mtx,
                                      const NDSRendererInputVertex *vtx,
                                      NDSRendererClipVertex20p12 *out)
 {
-    s64 x;
-    s64 y;
-    s64 z;
+    /* s32, not s64. The source fields are s16 and the matrix is s32, so every
+     * product here fits in 47 bits and SMULL computes it exactly; widening both
+     * operands first only forced the full 64x64 helper. Same value, one
+     * instruction instead of a call -- twelve times per vertex. */
+    s32 x;
+    s32 y;
+    s32 z;
 
     if ((mtx == NULL) || (vtx == NULL) || (out == NULL))
     {
